% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arabidopsis}
\alias{arabidopsis}
\title{Arabidopsis thaliana in the Iberian Peninsula}
\format{
## `arabidopsis`
A data frame with 301 rows and 15 columns:
\describe{
  \item{acronym}{Country name}
  \item{x, y}{Coordinates}
  \item{gc1, gc2, gc3, gc4}{Probability for belonging to each of the four genetic clusters: GC1, GC2, GC3 and GC4.}
  \item{bio1, bio2, bio3, bio4, bio8, bio12, bio15, bio18}{Bioclimatic variables corresponding with the definition in Worldclim}
  ...
}
}
\source{
<https://zenodo.org/record/2552025>
}
\usage{
arabidopsis
}
\description{
Collection of 301 accessions of the annual plant \textit{Arabidopsis thaliana} on the Iberian Peninsula. For each
accession, the probability of belonging to each of the 4 genetic clusters (GC) inferred in \cite{martinez-minaya2019}, namely, GC1, GC2, GC3 and GC4 represents the Compositional Data.
}
\keyword{datasets}
