% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bdiclcpomodel_stack.R
\name{Bdiclcpomodel_comp}
\alias{Bdiclcpomodel_comp}
\title{Selecting models based on DIC, WAIC, LCPO}
\usage{
Bdiclcpomodel_comp(
  variables,
  n = 256,
  shared.var = FALSE,
  spatial,
  spatial.short,
  friends = NULL,
  ...
)
}
\arguments{
\item{friends}{parameter for computing the CPO when we assume that some data points are friends.}

\item{variables:}{names of the terms to be included in the model selection process}

\item{n:}{number of models that we want to fit}

\item{...:}{more arguments for the inla function}
}
\value{
A list of three elements:
1: best modelS ordered by DIC
2: best models ordered by WAIC
3. best models ordered by LCPO
}
\description{
`Bdiclcpomodel_stack` is a function for selecting models based on DIC, WAIC, LCPO
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
