% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_mult.R
\name{dic.mult}
\alias{dic.mult}
\title{Computing dic for a Multivariate Gaussian likelihood}
\usage{
dic.mult(inf, y)
}
\arguments{
\item{inf}{a list with nsim elements. Each element of the list is a result of applying extract_lp_sigma.}

\item{y}{is a matrix in N x D with the response variable.}
}
\value{
e.dev, dev.e, p.eff, dic
}
\description{
`dic.mult` computes DIC for a Multivariate Gaussian likelihood
}
\author{
Joaquín Martínez-Minaya <\email{jmarminaya@gmail.com}>
}
