% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_mult.R
\name{dmnorm}
\alias{dmnorm}
\title{Computing multivariate Gaussian density}
\usage{
dmnorm(y, mu, Sigma, log1 = FALSE)
}
\arguments{
\item{y}{a vector (1 x d) or matrix (n x d) of multivariate observations}

\item{mu}{a vector (1 x d) or matrix (n x d) of multivariate observations}

\item{Sigma}{variance-covariance matrix (n x d)}

\item{log1}{TRUE if logaritmith transformation is used. Default is FALSE.}
}
\value{
Multivariate joint density Gaussian distribution
}
\description{
`dmnorm` computes the multivariate joint density Gaussian distribution
}
\author{
Joaquín Martínez-Minaya <\email{jmarminaya@gmail.com}>
}
