% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_mult.R
\name{extract_lp_sigma}
\alias{extract_lp_sigma}
\title{Extracting posterior distribution of the linear predictor a of the covariance matrix from an object obtained from inla.posterior.sample}
\usage{
extract_lp_sigma(xx)
}
\arguments{
\item{xx}{a resulting object of applying the function `inla.posterior.sample`}
}
\value{
A list with two objects: 1. Linear predictor substracting id.z. 2. Posterior Covariance matrix
}
\description{
`extract_lp_sigma` Extracts posterior distribution of the linear predictor a of the covariance matrix from an object obtained from inla.posterior.sample
}
\author{
Joaquín Martínez-Minaya <\email{jmarminaya@gmail.com}>
}
