% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygon_IP}
\alias{polygon_IP}
\title{Spatial Polygons of the Iberian Peninsula}
\format{
## `polygon_IP`
SpatialPolygonsDataFrame
\describe{
  \item{GID_0}{Country name}
  \item{NAME_0}{Country name}
  ...
}
}
\source{
GADM
}
\usage{
polygon_IP
}
\description{
Spatial polygons for the Iberian Peninsula
}
\keyword{datasets}
