% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygon_IP_low}
\alias{polygon_IP_low}
\title{Spatial Polygons of the Iberian Peninsula low resolution}
\format{
## `polygon_IP_low`
SpatialPolygonsDataFrame
\describe{
  \item{GID_0}{Country name}
  \item{NAME_0}{Country name}
  ...
}
}
\source{
GADM
}
\usage{
polygon_IP_low
}
\description{
Spatial polygons for the Iberian Peninsula low resolution
}
\keyword{datasets}
