% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_matrix.R
\name{precision_int}
\alias{precision_int}
\title{Computing structured matrix for space-time interactions in areal data
such as Q = tau*R}
\usage{
precision_int(term = c("besag", "rw2"), n, H)
}
\arguments{
\item{term}{a vector of two elements. In each one should include one of besag, rw1, rw2 or iid}

\item{n}{a vector with two elements with the number of points of each effect}

\item{H}{a list of two elements indicating inla graph, just required if term = "besag"}
}
\value{
list with the structured matrix of the term 1, term 2 and
the interaction between effect term1 and term2
}
\description{
`precision_matrix` Main function to compute R
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
