% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_matrix.R
\name{precision_matrix}
\alias{precision_matrix}
\title{Computing the structured matrix R, such as Q = tau*R}
\usage{
precision_matrix(term = "besag", n, H = NULL)
}
\arguments{
\item{term}{one of besag, rw1, rw2 or iid}

\item{n}{number of points}

\item{H}{inla graph, just required if term = "besag"}
}
\value{
Structured matrix R, such as Q = tau*R
}
\description{
`precision_matrix` Main function to compute R
}
\examples{
precision_matrix(term = "rw1", n = 10)
precision_matrix(term = "rw2", n = 15)
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
