% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raster_predict}
\alias{raster_predict}
\title{Raster for predicting spatial term and future predictions}
\format{
## `raster_predict`
raster_predict
\describe{
  ...
}
}
\source{
GADM
}
\usage{
raster_predict
}
\description{
Raster for predicting spatial term and future predictions
}
\keyword{datasets}
