% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_mult.R
\name{waic.mult}
\alias{waic.mult}
\title{Computing waic for a Multivariate Gaussian likelihood}
\usage{
waic.mult(inf, y)
}
\arguments{
\item{inf}{a list with nsim elements. Each element of the list is a result of applying extract_lp_sigma.}

\item{y}{is a matrix in N x D with the response variable.}
}
\value{
p.eff.waic, waic
}
\description{
`waic.mult` computes WAIC for a Multivariate Gaussian likelihood
}
\author{
Joaquín Martínez-Minaya <\email{jmarminaya@gmail.com}>
}
